refactor: extract nums/stats/technical to skills/lib/ shared layer
Move canonical implementations from invest-a-stock/scripts/lib/ to
skills/lib/, replacing originals with thin re-export shims.

- skills/lib/nums.py: safe_float, coalesce_field, fmt_amount
- skills/lib/stats.py: percentile_rank, calc_beta
- skills/lib/technical.py: compute, sma, rsi_series, sort_kline_asc, etc.
- Fix technical.py: from .stats import calc_beta → from stats import calc_beta
  (both now live in skills/lib/, no relative import needed)
- Shims re-export via sys.path bootstrap to skills/lib/ (already on path
  via existing _invest_path.py infrastructure in all skills)
- Explicitly export _ema, _rsi from technical shim for test compatibility

Zero changes needed to invest_path.py or cross-skill _invest_path.py
shims — skills/lib/ was already added to sys.path by all of them.
